Warning about run jet stream on existing projects.
I add red visible warning for peoples, that Jetstream will overwrite existing work, and it is meant to be install only on fresh installations.

I think the information about it is not clearly stated in the documentation. Unfortunately, I also did not understand the consequences of the installation by reading the documentation.