Add banner alerts example.
Added additional finding thanks to @jasonlbeggs for pointing this out on Twitter. I think this is a good addition especially for new starters to Jetstream.

I hope this can be added to the docs and thank you all for making such a great platform for us all to work on.