Point out that only team owners can manage team membership by default
When setting up Jetstream with teams, it is not totally clear what the 'admin' role means. One might think that an admin for a team would include team membership management, but it doesn't (it is, in fact, a whole different security model). This minor change to the docs brings that to light and proposes a way to let people change the policy.